% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\docType{data}
\name{qatar_maroon}
\alias{qatar_maroon}
\title{Qatar maroon color}
\format{
A character string containing a hex color code.
}
\source{
\url{https://en.wikipedia.org/wiki/Flag_of_Qatar}
}
\usage{
qatar_maroon
}
\description{
The official colors of the Qatari flag are white and Pantone 1955 C, or
"Qatar maroon." The hex representation of this color is \verb{#8A1538}.
}
\examples{
qatar_maroon

hist(qatarcars$length, breaks = 15, col = qatar_maroon, border = "white")

if (require("ggplot2")) {
  ggplot(qatarcars, aes(x = length)) +
    geom_histogram(bins = 15, fill = qatar_maroon, color = "white")
}

}
\keyword{datasets}
