% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{set_currency_label}
\alias{set_currency_label}
\title{Add a label attribute to an input value only if it already has a label attribute}
\usage{
set_currency_label(value, original, label)
}
\description{
This is useful with the currency conversion functions since the \code{price}
column in \code{qatarcars} has a label already ("Price (QAR)"), and if something
like \code{qar_to_usd()} is used on it, that label carries over.
}
\keyword{internal}
