% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_label}
\alias{set_label}
\title{Add a label attribute to an input value only if it already has a label attribute}
\usage{
set_label(value, original, label)
}
\description{
This is useful with conversion functions since columns in \code{qatarcars} have
labels already, and if a conversion function is used on them, that label
carries over.
}
\keyword{internal}
