% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qatarcars-package.R
\docType{package}
\name{qatarcars-package}
\alias{qatarcars-package}
\title{qatarcars: Data on Cars in Qatar}
\description{
Fuel economy, size, performance, and price data for cars in Qatar in 2025. Mirrors many of the columns in mtcars, but uses (1) non-US-centric makes and models, (2) 2025 prices, and (3) metric measurements, making it more appropriate for use as an example dataset outside the United States.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://profmusgrave.github.io/qatarcars/}
  \item \url{https://github.com/profmusgrave/qatarcars}
  \item Report bugs at \url{https://github.com/profmusgrave/qatarcars/issues}
}

}
\author{
\strong{Maintainer}: Paul Musgrave \email{rpm47@georgetown.edu} (\href{https://orcid.org/0000-0002-8984-4992}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Andrew Heiss \email{andrew@andrewheiss.com} (\href{https://orcid.org/0000-0002-3948-3914}{ORCID})
}

}
\keyword{internal}
